Expose a server api
Exposes a limited subset of the API that functions in Node.js.

Usage is:

```js
const L = require('leaflet/server');
```

Includes the following modules:
  * core
  * geo
  * geometry

The following modules are NOT included:
  * control
  * dom
  * layer
  * map
