+++ package.json
@@ -23,6 +23,7 @@
     "rollup": "0.51.8",
     "rollup-plugin-git-version": "0.2.1",
     "rollup-plugin-json": "^4.0.0",
+    "rollup-plugin-node-resolve": "^3.4.0",
     "sinon": "^7.3.2",
     "ssri": "^6.0.1",
     "uglify-js": "~3.5.10"
@@ -34,10 +35,14 @@
     "src",
     "!dist/leaflet.zip"
   ],
+  "browser": {
+    "./src/core/Browser.js": "./src/core/BrowserActual.js"
+  },
   "scripts": {
     "docs": "node ./build/docs.js",
     "pretest": "npm run lint && npm run lint-spec",
-    "test": "npm run test-nolint",
+    "test": "npm run test-node && npm run test-nolint",
+    "test-node": "mocha",
     "test-nolint": "karma start ./spec/karma.conf.js",
     "build": "npm run rollup && npm run uglify",
     "release": "./build/publish.sh",
@@ -48,6 +53,15 @@
     "watch": "rollup -w -c build/rollup-watch-config.js",
     "uglify": "uglifyjs dist/leaflet-src.js -c -m -o dist/leaflet.js --source-map filename=dist/leaflet.js.map --in-source-map dist/leaflet-src.js.map --source-map-url leaflet.js.map --comments",
     "integrity": "node ./build/integrity.js"
+  },
+  "mocha": {
+    "file": [
+      "./spec/ServerHelper.js",
+      "./spec/suites/SpecHelper.js"
+    ],
+    "spec": [
+      "./spec/suites/@(core|geometry|geo)/*.js"
+    ]
   },
   "eslintConfig": {
     "root": true,
@@ -97,5 +111,8 @@
     "gis",
     "map"
   ],
-  "license": "BSD-2-Clause"
+  "license": "BSD-2-Clause",
+  "dependencies": {
+    "esm": "^3.2.25"
+  }
 }
