+++ server.js
@@ -1,0 +1,2 @@
+module.exports = require('esm')(module)('./src/LeafletServer');
+
