+++ spec/ServerHelper.js
@@ -1,0 +1,4 @@
+global.L = require('../server');
+global.expect = require('expect.js');
+global.happen = require('happen');
+global.sinon = require('sinon');
