+++ spec/suites/core/GeneralSpec.js
@@ -1,5 +1,7 @@
 describe('General', function () {
-	describe('noConflict', function () {
+	it.skipInNode = L.Browser.node ? it.skip : it;
+
+	it.skipInNode('noConflict', function () {
 		var leaflet = L;
 
 		after(function () {
@@ -9,7 +11,7 @@ describe('General', function () {
 		expect(L.noConflict()).to.eql(leaflet);
 	});
 
-	describe('namespace extension', function () {
+	it('namespace extension', function () {
 		L.Util.foo = 'bar';
 		L.Foo = 'Bar';
