+++ src/LeafletServer.js
@@ -1,0 +1,24 @@
+// A subset of the API that will work in Node.
+
+import {version} from '../package.json';
+export {version};
+
+// control -- not supported
+
+// core
+export * from './core/index';
+
+// dom -- not supported
+
+// geometry
+export * from './geometry/index';
+
+// geo
+export * from './geo/index';
+
+// layer -- not supported
+
+// map -- not supported
+
+import {freeze} from './core/Util';
+Object.freeze = freeze;
